poker-backend-api/requirements.txt
treys
python-dotenv
uuid
boto3